* pitivi/playground.py: Added switchToTimeline() method. Don't seek to values lower than 0. * pitivi/timeline.py: Add default sources behaviours in compositions. Added setDefaultSource() and getDefaultSource() methods. * pitivi/ui/complexinterface.py: pixelToNs should return a long (to cope with 2**64 values). * pitivi/ui/complexlayer.py: Don't expand audio tracks at startup. * pitivi/ui/ruler.py: Allow seeking on the ruler. * pitivi/ui/sourcefactories.py: Use the source thumbnail as the icon pixbuf when doing drag and drop. * pitivi/ui/timelineobjects.py: Switch to Timeline in the playground if we drop something in it. * pitivi/ui/viewer.py: Use bilinear scaling method with videoscale. Use videosink pad to figure out caps modifications.
git-svn-id: svn+ssh://svn.gnome.org/svn/pitivi/trunk@801 d3729300-e425-0410-8a4c-d956edccc248

+++ pitivi/ui/complexinterface.py
@@ -75,7 +75,7 @@ class ZoomableWidgetInterface:
         """
         Returns the pixel equivalent in nanoseconds according to the zoomratio
         """
-        return int(pixel * gst.SECOND / self.getZoomRatio())
+        return long(pixel * gst.SECOND / self.getZoomRatio())
 
     def nsToPixel(self, duration):
         """
